+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Transistor method 'compare_channel_linearized()' to compare channel data with linearized data 
 
 ## [0.2.11] - 2021-08-07
 ### Added